refactor(main-home): #12 - 실시간 트렌드 섹션 컴포넌트로 분리
