--- app/components/realtime-trendcard.js
@@ -1,135 +1,0 @@
-import { Image, View } from 'react-native';
-import Svg, { ClipPath, Defs, Path, Image as SvgImage } from 'react-native-svg';
-
-const images = {
-  trendRank1Image: require('../../assets/images/trend_rank1.png'),
-  trendRank2Image: require('../../assets/images/trend_rank2.png'),
-  trendRank3Image: require('../../assets/images/trend_rank3.png'),
-  trendRank4Image: require('../../assets/images/trend_rank4.png'),
-};
-
-export default function TrendCard({
-  width,
-  height,
-  path,
-  children,
-  imageUrl,
-  contentStyle,
-}) {
-  const image = images[imageUrl];
-  const imageSource = Image.resolveAssetSource(image);
-
-  return (
-    <View style={{ overflow: 'hidden' }}>
-      <Svg width={width} height={height}>
-        <Defs>
-          <ClipPath id="clip">
-            <Path d={path} />
-          </ClipPath>
-        </Defs>
-        <SvgImage
-          href={imageSource}
-          width={width}
-          height={height}
-          preserveAspectRatio="xMidYMid slice"
-          clipPath="url(#clip)"
-        />
-      </Svg>
-      <View
-        style={[
-          {
-            position: 'absolute',
-            insets: 0,
-          },
-          contentStyle,
-        ]}
-      >
-        {children}
-      </View>
-    </View>
-  );
-}
-
-export const getFirstTrendCardPath = (width, height) => {
-  const cutX = width * 0.8;
-  const cutY = height * 0.47;
-  const r = 30;
-
-  return `
-    M0 ${r}
-    Q0 0 ${r} 0
-    H${width - r}
-    Q${width} 0 ${width} ${r}
-    V${cutY - r}
-    Q${width} ${cutY} ${width - r} ${cutY}
-    H${cutX + r}
-    Q${cutX} ${cutY} ${cutX} ${cutY + r}
-    V${height - r}
-    Q${cutX} ${height} ${cutX - r} ${height}
-    H${r}
-    Q0 ${height} 0 ${height - r}
-    Z
-  `;
-};
-
-export const getSecondTrendCardPath = (width, height) => {
-  const cutX = width * 0.75;
-  const cutY = height * 0.48;
-  const r = 30;
-
-  return `
-    M0 ${r}
-    Q0 0 ${r} 0
-    H${width - r}
-    Q${width} 0 ${width} ${r}
-    V${cutY - r}
-    Q${width} ${cutY} ${width - r} ${cutY}
-    H${cutX + r}
-    Q${cutX} ${cutY} ${cutX} ${cutY + r}
-    V${height - r}
-    Q${cutX} ${height} ${cutX - r} ${height}
-    H${r}
-    Q0 ${height} 0 ${height - r}
-    Z
-  `;
-};
-
-export const getThirdTrendCardPath = (width, height) => {
-  const cutX = width * 0.34;
-  const cutY = height * 0.52;
-  const r = 30;
-
-  return `
-    M${cutX + r} 0
-    H${width - r}
-    Q${width} 0 ${width} ${r}
-    V${height - r}
-    Q${width} ${height} ${width - r} ${height}
-    H${r}
-    Q0 ${height} 0 ${height - r}
-    V${cutY + r}
-    Q0 ${cutY} ${r} ${cutY}
-    H${cutX - r}
-    Q${cutX} ${cutY} ${cutX} ${cutY - r}
-    V${r}
-    Q${cutX} 0 ${cutX + r} 0
-    Z
-  `;
-};
-
-export const getFourthTrendCardPath = (width, height) => {
-  const r = 30;
-
-  return `
-    M${r} 0
-    H${width - r}
-    Q${width} 0 ${width} ${r}
-    V${height - r}
-    Q${width} ${height} ${width - r} ${height}
-    H${r}
-    Q0 ${height} 0 ${height - r}
-    V${r}
-    Q0 0 ${r} 0
-    Z
-  `;
-};
